Fix storage tests

+++ go/storage/rediscache_test.go
@@ -458,7 +458,7 @@ func TestRedisLogState(t *testing.T) {
 	log := &CertificateLog{
 		ShortURL:     "short_url/location",
 		MaxEntry:     123456789,
-		MaxTimestamp: time.Now().Unix(),
+		MaxTimestamp: uint64(time.Now().Unix()),
 	}
 
 	expectNilLogState(t, rc, log.ShortURL)
